Adding query yaml file and package refactoring.
